+++ packages/client/src/declarations/index.ts
@@ -77,6 +77,7 @@ import { EMPTY_STRING, SIGNATURE_KEYS } from '@client/utils/constants'
 import { ViewRecordQueries } from '@client/views/ViewRecord/query'
 import { UserDetails } from '@client/utils/userUtils'
 import { clearUnusedViewRecordCacheEntries } from '@client/utils/persistence'
+import { getReviewForm } from '@client/forms/register/review-selectors'
 
 const ARCHIVE_DECLARATION = 'DECLARATION/ARCHIVE'
 const SET_INITIAL_DECLARATION = 'DECLARATION/SET_INITIAL_DECLARATION'
@@ -1152,7 +1153,7 @@ function downloadDeclarationSuccess({
   store: IStoreState
   client: ApolloClient<{}>
 }): IDownloadDeclarationSuccess {
-  const form = getRegisterForm(store)
+  const form = getReviewForm(store)
 
   return {
     type: DOWNLOAD_DECLARATION_SUCCESS,