+++ packages/client/src/forms/index.ts
@@ -71,6 +71,7 @@ export const TIME = 'TIME'
 export const NID_VERIFICATION_BUTTON = 'NID_VERIFICATION_BUTTON'
 export const DIVIDER = 'DIVIDER'
 export const HEADING3 = 'HEADING3'
+export const SIGNATURE = 'SIGNATURE'
 
 export enum Sort {
   ASC = 'asc',
@@ -579,7 +580,6 @@ export interface INumberFormField extends IFormFieldBase {
   type: typeof NUMBER
   step?: number
   max?: number
-  inputFieldWidth?: string
   inputWidth?: number
   maxLength?: number
 }
@@ -712,6 +712,12 @@ export interface INidVerificationButton extends IFormFieldBase {
   labelForOffline: MessageDescriptor
 }
 
+export interface ISignatureFormField extends IFormFieldBase {
+  type: typeof SIGNATURE
+  maxSizeMb?: number
+  allowedFileFormats?: ('png' | 'jpg' | 'jpeg' | 'svg')[]
+}
+
 export type IFormField =
   | ITextFormField
   | ITelFormField
@@ -745,6 +751,7 @@ export type IFormField =
   | ITimeFormFIeld
   | INidVerificationButton
   | IDividerFormField
+  | ISignatureFormField
 
 export interface IPreviewGroup {
   id: string
@@ -995,7 +1002,7 @@ export interface Ii18nSelectOption {
 
 export interface Ii18nFormFieldBase {
   name: string
-  type: string
+  type: Ii18nFormField['type']
   label: string
   helperText?: string
   tooltip?: string
@@ -1078,7 +1085,6 @@ export interface Ii18nNumberFormField extends Ii18nFormFieldBase {
   type: typeof NUMBER
   step?: number
   max?: number
-  inputFieldWidth?: string
   inputWidth?: number
   maxLength?: number
 }
@@ -1199,6 +1205,13 @@ export interface Ii18nTimeFormField extends Ii18nFormFieldBase {
   type: typeof TIME
   ignorePlaceHolder?: boolean
 }
+
+export interface Ii18nSignatureField extends Ii18nFormFieldBase {
+  type: typeof SIGNATURE
+  maxSizeMb?: number
+  allowedFileFormats?: ('png' | 'jpg' | 'jpeg' | 'svg')[]
+}
+
 export type Ii18nFormField =
   | Ii18nTextFormField
   | Ii18nTelFormField
@@ -1230,6 +1243,7 @@ export type Ii18nFormField =
   | Ii18nNidVerificationButtonField
   | I18nDividerField
   | I18nHeading3Field
+  | Ii18nSignatureField
 
 export interface IFormSectionData {
   [key: string]: IFormFieldValue