+++ packages/client/src/views/RegisterForm/review/ReviewSection.tsx
@@ -31,6 +31,7 @@ import {
 import {
   IDeclaration,
   SUBMISSION_STATUS,
+  modifyDeclaration,
   writeDeclaration
 } from '@client/declarations'
 import { ReviewAction } from '@client/components/form/ReviewActionComponent'
@@ -73,10 +74,6 @@ import {
   HIDDEN
 } from '@client/forms'
 import { Event } from '@client/utils/gateway'
-import {
-  getBirthSection,
-  getRegisterForm
-} from '@client/forms/register/declaration-selectors'
 import {
   getConditionalActionsForField,
   getListOfLocations,
@@ -124,8 +121,8 @@ import {
 import { connect } from 'react-redux'
 import { ReviewHeader } from './ReviewHeader'
 import { IValidationResult } from '@client/utils/validate'
-import { DocumentListPreview } from '@client/components/form/DocumentUploadfield/DocumentListPreview'
-import { DocumentPreview } from '@client/components/form/DocumentUploadfield/DocumentPreview'
+import { DocumentListPreview } from '@client/components/form/DocumentUploadField/DocumentListPreview'
+import { DocumentPreview } from '@client/components/form/DocumentUploadField/DocumentPreview'
 import { generateLocations } from '@client/utils/locationUtils'
 import {
   addressFieldNames,
@@ -136,13 +133,10 @@ import {
 import { ListReview } from '@opencrvs/components/lib/ListReview'
 import { DuplicateWarning } from '@client/views/Duplicates/DuplicateWarning'
 import { VerificationButton } from '@opencrvs/components/lib/VerificationButton'
-import {
-  SignatureGenerator,
-  SignatureInputProps
-} from '@client/views/RegisterForm/review/SignatureGenerator'
 import { DuplicateForm } from '@client/views/RegisterForm/duplicate/DuplicateForm'
 import { Button } from '@opencrvs/components/lib/Button'
 import { UserDetails } from '@client/utils/userUtils'
+import { FormFieldGenerator } from '@client/components/form'
 
 const Deleted = styled.del`
   color: ${({ theme }) => theme.colors.negative};
@@ -273,10 +267,11 @@ type onChangeReviewForm = (
 
 interface IProps {
   draft: IDeclaration
-  registerForm: { [key: string]: IForm }
+  form: IForm
   pageRoute: string
   rejectDeclarationClickEvent?: () => void
   goToPageGroup: typeof goToPageGroup
+  modifyDeclaration: typeof modifyDeclaration
   submitClickEvent: (
     declaration: IDeclaration,
     submissionStatus: string,
@@ -633,14 +628,6 @@ export const getErrorsOnFieldsBySection = (
   }, {})
 }
 
-enum SignatureSectionType {
-  INFORMANT_SIGNATURE = 'informantsSignature',
-  BRIDE_SIGNATURE = 'brideSignature',
-  GROOM_SIGNATURE = 'groomSignature',
-  WITNESS_ONE_SIGNATURE = 'witnessOneSignature',
-  WITNESS_TWO_SIGNATURE = 'witnessTwoSignature'
-}
-
 class ReviewSectionComp extends React.Component<FullProps, State> {
   hasChangesBeenMade = false
 
@@ -675,8 +662,8 @@ class ReviewSectionComp extends React.Component<FullProps, State> {
   }
 
   getViewableSection = (registerForm: IForm): IFormSection[] => {
-    const sections = registerForm.sections.filter(
-      ({ viewType }) => viewType === 'form' || viewType === 'hidden'
+    const sections = registerForm.sections.filter(({ viewType }) =>
+      ['form', 'hidden', 'preview', 'review'].includes(viewType)
     )
 
     return this.getVisibleSections(sections)
@@ -684,9 +671,9 @@ class ReviewSectionComp extends React.Component<FullProps, State> {
 
   getLabelForDoc = (docForWhom: string, docType: string) => {
     const { intl } = this.props
-    const documentSection = this.props.registerForm[
-      this.props.draft.event
-    ].sections.find((section) => section.id === 'documents')
+    const documentSection = this.props.form.sections.find(
+      (section) => section.id === 'documents'
+    )
     const docSectionFields = documentSection && documentSection.groups[0].fields
     const docFieldsWithOptions =
       docSectionFields &&
@@ -1440,7 +1427,7 @@ class ReviewSectionComp extends React.Component<FullProps, State> {
       overriddenField,
       'reviewOverrides.residingSection'
     )
-    const residingSection = this.props.registerForm.death.sections.find(
+    const residingSection = this.props.form.sections.find(
       (section) => section.id === residingSectionId
     ) as IFormSection
 
@@ -1542,8 +1529,8 @@ class ReviewSectionComp extends React.Component<FullProps, State> {
   }
 
   isLastNameFirst = () => {
-    const { registerForm, draft: declaration } = this.props
-    const fields = registerForm[declaration.event].sections.find((section) =>
+    const { form, draft: declaration } = this.props
+    const fields = form.sections.find((section) =>
       declaration.event === Event.Birth
         ? section.id === 'child'
         : section.id === 'deceased'
@@ -1655,7 +1642,7 @@ class ReviewSectionComp extends React.Component<FullProps, State> {
     const {
       intl,
       draft: declaration,
-      registerForm,
+      form,
       rejectDeclarationClickEvent,
       submitClickEvent,
       registrationSection,
@@ -1669,7 +1656,7 @@ class ReviewSectionComp extends React.Component<FullProps, State> {
     const isDuplicate = Boolean(declaration.duplicates?.length)
     const formSections =
       viewRecord || isDuplicate
-        ? this.getViewableSection(registerForm[event]).map((section) => {
+        ? this.getViewableSection(form).map((section) => {
             return {
               ...section,
               groups: section.groups.map((group) => {
@@ -1680,7 +1667,7 @@ class ReviewSectionComp extends React.Component<FullProps, State> {
               })
             }
           })
-        : this.getViewableSection(registerForm[event])
+        : this.getViewableSection(form)
     const errorsOnFields = getErrorsOnFieldsBySection(
       formSections,
       offlineCountryConfiguration,
@@ -1693,31 +1680,10 @@ class ReviewSectionComp extends React.Component<FullProps, State> {
       true
     )
 
-    const isSignatureMissing = () => {
-      if (isCorrection(declaration)) {
-        return false
-      } else {
-        if (event === Event.Birth || event === Event.Death) {
-          return (
-            offlineCountryConfiguration.config.FEATURES
-              .INFORMANT_SIGNATURE_REQUIRED &&
-            !declaration.data.registration?.informantsSignature
-          )
-        } else if (event === Event.Marriage) {
-          return (
-            !declaration.data.registration?.groomSignature ||
-            !declaration.data.registration?.brideSignature ||
-            !declaration.data.registration?.witnessOneSignature ||
-            !declaration.data.registration?.witnessTwoSignature
-          )
-        }
-      }
-    }
-
     const isComplete =
       flatten(Object.values(errorsOnFields).map(Object.values)).filter(
         (errors) => errors.errors.length > 0
-      ).length === 0 && !isSignatureMissing()
+      ).length === 0
     const hasValidationErrors =
       flatten(Object.values(badInputErrors).map(Object.values)).filter(
         (errors) => errors.errors.length > 0
@@ -1744,112 +1710,31 @@ class ReviewSectionComp extends React.Component<FullProps, State> {
     )
     const draft = this.isDraft()
     const transformedSectionData = this.transformSectionData(
-      formSections.filter(({ id }) => id !== 'documents'),
+      formSections.filter(
+        ({ id, viewType }) =>
+          id !== 'documents' && !['preview', 'review'].includes(viewType)
+      ),
       errorsOnFields,
       offlineCountryConfiguration,
       declaration
     )
     const totalFileSizeExceeded = isFileSizeExceeded(declaration)
 
-    const generateSignatureProps = (
-      sectionType: SignatureSectionType,
-      id: string,
-      value: string,
-      inputLabel: string,
-      isRequired: boolean
-    ): SignatureInputProps => {
-      return {
-        id: id,
-        onChange: (value: string) => {
-          this.props.onChangeReviewForm &&
-            this.props.onChangeReviewForm(
-              {
-                [sectionType]: value,
-                [sectionType + 'URI']: ''
-              },
-              registrationSection,
-              declaration
-            )
-        },
-        value: value,
-        label: inputLabel,
-        isRequired: isRequired
-      }
-    }
-
-    const getSignature = (eventType: Event) => {
-      if ([Event.Birth, Event.Death].includes(eventType)) {
-        if (!window.config.FEATURES.INFORMANT_SIGNATURE) {
-          return <></>
+    const modifyDeclaration = (
+      sectionData: IFormSectionData,
+      section: IFormSection,
+      declaration: IDeclaration
+    ) => {
+      this.props.modifyDeclaration({
+        ...declaration,
+        data: {
+          ...declaration.data,
+          [section.id]: {
+            ...declaration.data[section.id],
+            ...sectionData
+          }
         }
-        const informantsSignatureInputPros = generateSignatureProps(
-          SignatureSectionType.INFORMANT_SIGNATURE,
-          'informants_signature',
-          declaration.data.registration?.informantsSignature as string,
-          intl.formatMessage(messages.informantsSignature),
-          window.config.FEATURES.INFORMANT_SIGNATURE_REQUIRED
-        )
-        return (
-          <>
-            <Accordion
-              name="signatures"
-              label="Signatures"
-              labelForHideAction={intl.formatMessage(messages.hideLabel)}
-              labelForShowAction={intl.formatMessage(messages.showLabel)}
-              expand={true}
-            >
-              <SignatureGenerator
-                description={intl.formatMessage(messages.signatureDescription)}
-                {...informantsSignatureInputPros}
-              />
-            </Accordion>
-          </>
-        )
-      } else if ([Event.Marriage].includes(eventType)) {
-        const brideSignatureInputPros = generateSignatureProps(
-          SignatureSectionType.BRIDE_SIGNATURE,
-          'bride_signature',
-          declaration.data.registration?.brideSignature as string,
-          intl.formatMessage(messages.brideSignature),
-          true
-        )
-
-        const groomSignatureInputPros = generateSignatureProps(
-          SignatureSectionType.GROOM_SIGNATURE,
-          'groom_signature',
-          declaration.data.registration?.groomSignature as string,
-          intl.formatMessage(messages.groomSignature),
-          true
-        )
-
-        const witnessOneSignatureInputPros = generateSignatureProps(
-          SignatureSectionType.WITNESS_ONE_SIGNATURE,
-          'witness_one_signature',
-          declaration.data.registration?.witnessOneSignature as string,
-          intl.formatMessage(messages.witnessOneSignature),
-          true
-        )
-
-        const witnessTwoSignatureInputPros = generateSignatureProps(
-          SignatureSectionType.WITNESS_TWO_SIGNATURE,
-          'witness_two_signature',
-          declaration.data.registration?.witnessTwoSignature as string,
-          intl.formatMessage(messages.witnessTwoSignature),
-          true
-        )
-
-        return (
-          <>
-            <Text id="terms" element="p" variant="reg16">
-              {intl.formatMessage(messages.terms)}
-            </Text>
-            <SignatureGenerator {...groomSignatureInputPros} />
-            <SignatureGenerator {...brideSignatureInputPros} />
-            <SignatureGenerator {...witnessOneSignatureInputPros} />
-            <SignatureGenerator {...witnessTwoSignatureInputPros} />
-          </>
-        )
-      }
+      })
     }
 
     const options = this.prepSectionDocOptions(declaration)
@@ -1858,6 +1743,16 @@ class ReviewSectionComp extends React.Component<FullProps, State> {
         this.isVisibleField(field, documentsSection)
       ).length
     )
+
+    const reviewSection = formSections.find(({ viewType }) =>
+      ['preview', 'review'].includes(viewType)
+    )
+
+    if (!reviewSection) {
+      throw new Error(
+        'The form definition should always contain a review or preview section'
+      )
+    }
 
     return (
       <Wrapper>
@@ -2024,10 +1919,20 @@ class ReviewSectionComp extends React.Component<FullProps, State> {
                     </Accordion>
                   )}
 
-                  {!isCorrection(declaration) &&
-                    !isDuplicate &&
-                    !viewRecord &&
-                    getSignature(declaration?.event as Event)}
+                  <FormFieldGenerator
+                    id={reviewSection.id}
+                    key={reviewSection.id}
+                    onChange={(values) => {
+                      modifyDeclaration(values, reviewSection, declaration)
+                    }}
+                    setAllFieldsDirty={false}
+                    fields={reviewSection.groups[0].fields}
+                    draftData={declaration.data}
+                    initialValues={
+                      declaration.data.preview ?? declaration.data.review
+                    }
+                  />
+
                   {totalFileSizeExceeded && (
                     <StyledAlert type="warning">
                       {intl.formatMessage(
@@ -2185,15 +2090,22 @@ function fieldToReadOnlyFields(field: IFormField): IFormField {
   }
   return readyOnlyField
 }
-
 export const ReviewSection = connect(
-  (state: IStoreState) => ({
-    registerForm: getRegisterForm(state),
-    registrationSection: getBirthSection(state, 'registration'),
-    documentsSection: getBirthSection(state, 'documents'),
-    scope: getScope(state),
-    offlineCountryConfiguration: getOfflineData(state),
-    language: getLanguage(state)
-  }),
-  { goToPageGroup, writeDeclaration }
+  (state: IStoreState, { form }: { form: IForm }) => {
+    const registrationSection = form.sections.find(
+      ({ id }) => id === 'registration'
+    )
+    const documentsSection = form.sections.find(({ id }) => id === 'documents')
+    if (!registrationSection || !documentsSection) {
+      throw new Error('"registration" & "documents" are required sections')
+    }
+    return {
+      registrationSection,
+      documentsSection,
+      scope: getScope(state),
+      offlineCountryConfiguration: getOfflineData(state),
+      language: getLanguage(state)
+    }
+  },
+  { goToPageGroup, writeDeclaration, modifyDeclaration }
 )(injectIntl(ReviewSectionComp))