+++ packages/client/src/views/SysAdmin/Performance/reports/operational/StatusWiseDeclarationCountView.tsx
@@ -9,7 +9,7 @@
  * Copyright (C) The OpenCRVS Authors located at https://github.com/opencrvs/opencrvs-core/blob/master/AUTHORS.
  */
 import { messages as performanceMessages } from '@client/i18n/messages/views/performance'
-import { RegStatus } from '@client/utils/gateway'
+import { RegStatus, Event } from '@client/utils/gateway'
 
 import {
   Description,
@@ -25,12 +25,14 @@ import {
 } from 'react-intl'
 import styled from 'styled-components'
 import { checkExternalValidationStatus } from '@client/views/SysAdmin/Team/utils'
-import { Event } from '@client/utils/gateway'
 
 type Props = WrappedComponentProps & BaseProps
 
 export type IStatusMapping = {
-  [status in Exclude<RegStatus, RegStatus.DeclarationUpdated>]: { labelDescriptor: MessageDescriptor; color: string }
+  [status in Exclude<RegStatus, RegStatus.DeclarationUpdated>]: {
+    labelDescriptor: MessageDescriptor
+    color: string
+  }
 }
 
 interface BaseProps {
@@ -77,13 +79,20 @@ class StatusWiseDeclarationCountViewComponent extends React.Component<
                   <ProgressBar
                     id={`${statusCount.status.toLowerCase()}-${index}`}
                     title={intl.formatMessage(
-                      statusMapping![statusCount.status as keyof IStatusMapping].labelDescriptor
+                      statusMapping![statusCount.status as keyof IStatusMapping]
+                        .labelDescriptor
                     )}
-                    color={statusMapping![statusCount?.status as keyof IStatusMapping].color}
+                    color={
+                      statusMapping![
+                        statusCount?.status as keyof IStatusMapping
+                      ].color
+                    }
                     totalPoints={total}
                     disabled={!isAccessibleOffice}
                     onClick={() =>
-                      this.props.onClickStatusDetails(statusCount.status as keyof IStatusMapping)
+                      this.props.onClickStatusDetails(
+                        statusCount.status as keyof IStatusMapping
+                      )
                     }
                     currentPoints={statusCount.count}
                   />